update browser; refactor; add login and symlink on startup

+++ .profile
@@ -3,7 +3,7 @@ export EDITOR="nvim"
 export VISUAL="nvim"
 export SUDO_EDITOR="nvim"
 export TERMINAL="st"
-export BROWSER="chromium"
+export BROWSER="google-chrome-unstable"
 export READER="zathura"
 export FILE="lf"
 export STATUSBAR="dwmbar"
@@ -11,7 +11,6 @@ export XIVIEWER="sxiv"
 
 export COLORTERM="truecolor"
 
-export INPUTRC="$HOME/.config/inputrc"
 export ZDOTDIR="$HOME/.config/zsh"
 export KEYTIMEOUT=1
 
@@ -37,9 +36,12 @@ export zathurarc="$HOME/.config/zathura/zathurarc"
 export pistolrc="$HOME/.config/pistol/pistol.conf"
 export dunst="$HOME/.config/dunst/dunstrc"
 
+export NPM_PACKAGES="${HOME}/.npm-packages"
+export prefix="${HOME}/.npm-packages"
+export NODE_PATH="$NPM_PACKAGES/lib/node_modules:$NODE_PATH"
+
 # PATH
 export MANPATH="${MANPATH-$(manpath)}:$NPM_PACKAGES/share/man"
-export NPM_PACKAGES="${HOME}/.npm-packages"
 export PATH="$NPM_PACKAGES/bin:$PATH" # NPM without sudo
 export PATH="$HOME/.basher/bin:$PATH" # Basher
 export PATH="$HOME/.cargo/bin:$PATH" # Cargo
@@ -60,4 +62,5 @@ export LESS=-r
 eval "$(dircolors $HOME/.config/LS_COLORS)"
 
 mkdir -p /tmp/downloads
-ln -s $HOME/media/articles /tmp/downloads/articles
+[[ ! -L /tmp/downloads/articles ]] && ln -s $HOME/media/articles /tmp/downloads/articles
+[ "$(tty)" = "/dev/tty2" ] && ! pgrep -x Xorg >/dev/null && exec startx
